backend/requirements.txt
# API
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.3.0

# OpenAI
openai>=1.3.0

# CORS
python-multipart>=0.0.6